refactor(lib): simplify methods and refactor into separate files
