+++ tasks/lib/get-previous-tag.js
@@ -1,0 +1,33 @@
+'use strict';
+
+var debug = require('debug')('changelog:getPreviousTag');
+var child = require('child_process');
+var q = require('q');
+
+function cmdDone(deferred, code, stdout, stderr) {
+  debug('returning from git tag');
+  if (code) {
+    deferred.reject();
+  } else {
+    deferred.resolve(stdout.replace('\n', ''));
+  }
+}
+
+function getPreviousTag() {
+  debug('getting previous tag');
+  var deferred = q.defer();
+
+  if (this.options.tag) {
+    deferred.resolve(this.options.tag);
+  } else if (this.options.tag === false) {
+    deferred.resolve(false);
+  } else {
+    //IF we dont find a previous tag, we get all the commits from the beggining - The bigbang of the code
+    debug('calling git tag command');
+    child.exec(this.cmd.gitTag, cmdDone.bind(null, deferred));
+  }
+
+  return deferred.promise;
+}
+
+module.exports = getPreviousTag;
